Remove TypeUtil changes from previous attempt
