+++ api/src/main/java/org/apache/iceberg/types/AssignFreshIds.java
@@ -18,13 +18,19 @@
  */
 package org.apache.iceberg.types;
 
+import java.util.Iterator;
+import java.util.List;
+import java.util.function.Supplier;
 import org.apache.iceberg.Schema;
+import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 
-class AssignFreshIds extends BaseAssignIds {
+class AssignFreshIds extends TypeUtil.CustomOrderSchemaVisitor<Type> {
+  private final Schema visitingSchema;
   private final Schema baseSchema;
   private final TypeUtil.NextID nextId;
 
   AssignFreshIds(TypeUtil.NextID nextId) {
+    this.visitingSchema = null;
     this.baseSchema = null;
     this.nextId = nextId;
   }
@@ -37,13 +43,12 @@ class AssignFreshIds extends BaseAssignIds {
    * @param nextId new id assigner
    */
   AssignFreshIds(Schema visitingSchema, Schema baseSchema, TypeUtil.NextID nextId) {
-    super(visitingSchema);
+    this.visitingSchema = visitingSchema;
     this.baseSchema = baseSchema;
     this.nextId = nextId;
   }
 
-  @Override
-  protected int idFor(String fullName) {
+  private int idFor(String fullName) {
     if (baseSchema != null && fullName != null) {
       Types.NestedField field = baseSchema.findField(fullName);
       if (field != null) {
@@ -52,5 +57,75 @@ class AssignFreshIds extends BaseAssignIds {
     }
 
     return nextId.get();
+  }
+
+  private String name(int id) {
+    if (visitingSchema != null) {
+      return visitingSchema.findColumnName(id);
+    }
+
+    return null;
+  }
+
+  @Override
+  public Type schema(Schema schema, Supplier<Type> future) {
+    return future.get();
+  }
+
+  @Override
+  public Type struct(Types.StructType struct, Iterable<Type> futures) {
+    List<Types.NestedField> fields = struct.fields();
+    int length = struct.fields().size();
+
+    // assign IDs for this struct's fields first
+    List<Integer> newIds = Lists.newArrayListWithExpectedSize(length);
+    for (int i = 0; i < length; i += 1) {
+      newIds.add(idFor(name(fields.get(i).fieldId())));
+    }
+
+    List<Types.NestedField> newFields = Lists.newArrayListWithExpectedSize(length);
+    Iterator<Type> types = futures.iterator();
+    for (int i = 0; i < length; i += 1) {
+      Types.NestedField field = fields.get(i);
+      Type type = types.next();
+      if (field.isOptional()) {
+        newFields.add(Types.NestedField.optional(newIds.get(i), field.name(), type, field.doc()));
+      } else {
+        newFields.add(Types.NestedField.required(newIds.get(i), field.name(), type, field.doc()));
+      }
+    }
+
+    return Types.StructType.of(newFields);
+  }
+
+  @Override
+  public Type field(Types.NestedField field, Supplier<Type> future) {
+    return future.get();
+  }
+
+  @Override
+  public Type list(Types.ListType list, Supplier<Type> future) {
+    int newId = idFor(name(list.elementId()));
+    if (list.isElementOptional()) {
+      return Types.ListType.ofOptional(newId, future.get());
+    } else {
+      return Types.ListType.ofRequired(newId, future.get());
+    }
+  }
+
+  @Override
+  public Type map(Types.MapType map, Supplier<Type> keyFuture, Supplier<Type> valueFuture) {
+    int newKeyId = idFor(name(map.keyId()));
+    int newValueId = idFor(name(map.valueId()));
+    if (map.isValueOptional()) {
+      return Types.MapType.ofOptional(newKeyId, newValueId, keyFuture.get(), valueFuture.get());
+    } else {
+      return Types.MapType.ofRequired(newKeyId, newValueId, keyFuture.get(), valueFuture.get());
+    }
+  }
+
+  @Override
+  public Type primitive(Type.PrimitiveType primitive) {
+    return primitive;
   }
 }
